requirements.txt
fastapi
uvicorn[standard]
httpx
asyncpg
yt-dlp
jinja2>=3.1.2
psycopg2-binary
alembic
apscheduler
pydantic
starlette
camoufox
aiogram
aiohttp
python-dotenv
redis==5.2.1
langdetect
SQLAlchemy